requirements.txt
streamlit>=1.24.0
langflow>=0.5.0
langchain>=0.0.300
openai>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0